refactor: allocator agent card naming structure
