+++ apps/torus-allocator/src/app/(expanded-pages)/agent/[slug]/components/expanded-view-socials.tsx
@@ -1,4 +1,4 @@
-import { buildSocials } from "~/app/_components/agent-item-card/components/socials-info";
+import { buildSocials } from "~/app/_components/agent-card/components/agent-card-socials-info";
 import Link from "next/link";
 
 interface ExpandedViewSocialsProps {
