+++ apps/torus-allocator/src/app/_components/agent-card/agent-card.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { AgentCardContent } from "./components/agent-card-content";
 import { AgentCardFooter } from "./components/agent-card-footer";
 import { AgentCardHeader } from "./components/agent-card-header";
-import { CardHoverEffect } from "./components/card-hover-effect";
+import { CardHoverEffect } from "./components/agent-card-hover-effect";
 
 interface AgentCardProps {
   id: number;
@@ -15,7 +15,7 @@ interface AgentCardProps {
   weightFactor: number | null;
 }
 
-export function AgentItemCard(props: Readonly<AgentCardProps>) {
+export function AgentCard(props: Readonly<AgentCardProps>) {
   return (
     <Card
       className="to-background group relative border bg-gradient-to-tr from-zinc-900 transition
